rename GREYConfigurationApi to GREYConfiguration
