+++ detox/src/ios/earlgreyapi/GREYConfiguration.js
