+++ generation/index.js
@@ -16,7 +16,7 @@ const iosFiles = {
 	"../detox/ios/Detox/GREYConfiguration+Detox.h":
 		"../detox/src/ios/earlgreyapi/GREYConfigurationDetox.js",
 	"../detox/ios/EarlGrey/EarlGrey/Common/GREYConfiguration.h":
-		"../detox/src/ios/earlgreyapi/GREYConfigurationApi.js"
+		"../detox/src/ios/earlgreyapi/GREYConfiguration.js"
 };
 
 generateIOSAdapters(iosFiles);
